+++ CHANGELOG.md
@@ -123,3 +123,7 @@
 ## v2.0.2
 * Fixed bug when legends is activated and no wms is on the map
 * Fixed padding when watermark subtitle is largen than title
+
+## v2.0.3
+* Improved attributions functionality. Now extracts the data from the layers and not form the ui control
+* Count loaded tiles and images in all layers (not only wms)